move http server types to their own place
